Removed WorldObject
It doesn't make sense to have Block and Item extend some arbitrary WorldObject. It makes a lot of sense for Block to extend Item, since every Block has a corresponding Item that it builds on.
