+++ src/main/java/us/timinc/jsonifycraft/description/JsonDescription.java
@@ -3,6 +3,7 @@ package us.timinc.jsonifycraft.description;
 import java.util.Arrays;
 
 public class JsonDescription {
+    public String name;
     public String type;
     public String[] flags = {};
